Atualização BackEnd

+++ Back/app/api/contato.js
@@ -47,113 +47,134 @@ module.exports = function(app){
 
     //Lista as informações de uma Pessoa e seus Telefones.
     api.listaContatoInfo = (req, res) => {
-        const connection = app.conexao.conexaoBD();
-        const contatoDAO = new app.infra.ContatoDAO(connection);
         const { cod_pessoa } = req.params;
 
-        contatoDAO.listarContatoInfo(cod_pessoa, (erro, resultado) => {
-            erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
-        });
+        if(cod_pessoa){
+            const connection = app.conexao.conexaoBD();
+            const contatoDAO = new app.infra.ContatoDAO(connection);
 
-        connection.end();
+            contatoDAO.listarContatoInfo(cod_pessoa, (erro, resultado) => {
+                erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
+            });
+
+            connection.end();
+        } else { res.status(400).send(app.api.erroPadrao()); }
     };
 
     //Lista os Contatos de uma Instituicao Responsavel.
     api.listaContatoInstituicao = (req, res) => {
-        const connection = app.conexao.conexaoBD();
-        const contatoDAO = new app.infra.ContatoDAO(connection);
         const { cod_instituicao } = req.params;
 
-        contatoDAO.listarContatoInstituicao(cod_instituicao, (erro, resultado) => {
-            erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
-        });
+        if(cod_instituicao){
+            const connection = app.conexao.conexaoBD();
+            const contatoDAO = new app.infra.ContatoDAO(connection);
 
-        connection.end();
+            contatoDAO.listarContatoInstituicao(cod_instituicao, (erro, resultado) => {
+                erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
+            });
+
+            connection.end();
+        } else { res.status(400).send(app.api.erroPadrao()); }
     };
 
     //Lista os Contatos de uma Empresa.
     api.listaContatoEmpresa = (req, res) => {
-        const connection = app.conexao.conexaoBD();
-        const contatoDAO = new app.infra.ContatoDAO(connection);
         const { cnpj_empresa } = req.params;
 
-        contatoDAO.listarContatoEmpresa(cnpj_empresa, (erro, resultado) => {
-            erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
-        });
+        if(cnpj_empresa){
+            const connection = app.conexao.conexaoBD();
+            const contatoDAO = new app.infra.ContatoDAO(connection);
 
-        connection.end();
+            contatoDAO.listarContatoEmpresa(cnpj_empresa, (erro, resultado) => {
+                erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
+            });
+
+            connection.end();
+        } else { res.status(400).send(app.api.erroPadrao()); }
     };
 
     //Lista os Contatos de um Pontos de Presença.
     api.listaContatoPonto = (req, res) => {
-        const connection = app.conexao.conexaoBD();
-        const contatoDAO = new app.infra.ContatoDAO(connection);
         const { cod_gesac } = req.params;
 
-        contatoDAO.listarContatoPonto(cod_gesac, (erro, resultado) => {
-            erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
-        });
+        if(cod_gesac){
+            const connection = app.conexao.conexaoBD();
+            const contatoDAO = new app.infra.ContatoDAO(connection);
 
-        connection.end();
+            contatoDAO.listarContatoPonto(cod_gesac, (erro, resultado) => {
+                erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado);
+            });
+
+            connection.end();
+        } else { res.status(400).send(app.api.erroPadrao()); }
     };
 
     //Lista as informações de um Contatos.
     api.listaContatoDados = (req, res) => {
-        const connection = app.conexao.conexaoBD();
-        const contatoDAO = new app.infra.ContatoDAO(connection);
         const { cod_contato } = req.params;
 
-        contatoDAO.listarContatoDados(cod_contato, (erro, resultado) => {
-            erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado[0]);
-        });
+        if(cod_contato){
+            const connection = app.conexao.conexaoBD();
+            const contatoDAO = new app.infra.ContatoDAO(connection);
 
-        connection.end();
+            contatoDAO.listarContatoDados(cod_contato, (erro, resultado) => {
+                erro ? (console.log(erro), res.status(500).send(app.api.erroPadrao())) : res.status(200).json(resultado[0]);
+            });
+
+            connection.end();
+        } else { res.status(400).send(app.api.erroPadrao()); }
     };
 
     //Atualiza os dados de um Contato.
     api.editaContato = (req, res) => {
-        const knex = app.conexao.conexaoBDKnex();
         const { cod_contato } = req.params;
-        const { cargo, obs, cod_pessoa, nome } = req.body;
 
-        const contato = { cargo, obs};
+        if(cod_contato){
+            const knex = app.conexao.conexaoBDKnex();
+            const { cargo, obs, cod_pessoa, nome } = req.body;
 
-        knex('contato').where('cod_contato', cod_contato).update(contato)
-            .then(resultado => {
-                const pessoa = { nome };
+            const contato = { cargo, obs};
 
-                knex('pessoa').where('cod_pessoa', cod_pessoa).update(pessoa)
-                    .then(result => {
-                        knex.destroy();
-                        res.status(200).end();
-                    })
-            })
-            .catch(erro => {
-                console.log(erro);
-                knex.destroy();
-                res.status(500).send(app.api.erroPadrao());
-            });
+            knex('contato').where('cod_contato', cod_contato).update(contato)
+                .then(resultado => {
+                    const pessoa = { nome };
+
+                    knex('pessoa').where('cod_pessoa', cod_pessoa).update(pessoa)
+                        .then(result => {
+                            knex.destroy();
+                            res.status(200).end();
+                        })
+                })
+                .catch(erro => {
+                    console.log(erro);
+                    knex.destroy();
+                    res.status(500).send(app.api.erroPadrao());
+                });
+        } else { res.status(400).send(app.api.erroPadrao()); }
     }
 
     //Apaga um Contato com base no cod_contato.
     api.apagaContato = (req, res) => {
-        const knex = app.conexao.conexaoBDKnex();
         const { cod_contato } = req.params;
 
-        knex('contato').where('cod_contato', cod_contato).delete()
-            .then(resultado => {
-                knex.destroy();
-                res.status(200).end();
-            })
-            .catch(erro => {
-                console.log(erro);
-                knex.destroy();
-                if(erro.errno == 1451){
-                    res.status(500).send('Este contato não pode ser apagado pois existem outras informações associadas a ele.');
-                } else {
-                    res.status(500).send(app.api.erroPadrao());
-                }
-            });
+        if(cod_contato){
+            const knex = app.conexao.conexaoBDKnex();
+
+            knex('contato').where('cod_contato', cod_contato).delete()
+                .then(resultado => {
+                    knex.destroy();
+                    res.status(200).end();
+                })
+                .catch(erro => {
+                    console.log(erro);
+                    knex.destroy();
+                    if(erro.errno == 1451){
+                        res.status(500).send('Este contato não pode ser apagado pois existem outras informações associadas a ele.');
+                    } else {
+                        res.status(500).send(app.api.erroPadrao());
+                    }
+                });
+        } else { res.status(400).send(app.api.erroPadrao()); }
     }
 
 
@@ -177,37 +198,43 @@ module.exports = function(app){
 
     //Atualiza os dados de um Telefone.
     api.editaTelefone = (req, res) => {
-        const knex = app.conexao.conexaoBDKnex();
         const { cod_telefone, cod_pessoa } = req.params;
-        const telefone = req.body;
 
-        knex('telefone').where('cod_telefone', cod_telefone).andWhere('cod_pessoa', cod_pessoa).update(telefone)
-            .then(resultado => {
-                knex.destroy();
-                res.status(200).end();
-            })
-            .catch(erro => {
-                console.log(erro);
-                knex.destroy();
-                res.status(500).send(app.api.erroPadrao());
-            });
+        if(cod_telefone && cod_pessoa){
+            const knex = app.conexao.conexaoBDKnex();
+            const telefone = req.body;
+
+            knex('telefone').where('cod_telefone', cod_telefone).andWhere('cod_pessoa', cod_pessoa).update(telefone)
+                .then(resultado => {
+                    knex.destroy();
+                    res.status(200).end();
+                })
+                .catch(erro => {
+                    console.log(erro);
+                    knex.destroy();
+                    res.status(500).send(app.api.erroPadrao());
+                });
+        } else { res.status(400).send(app.api.erroPadrao()); }
     }
 
     //Apaga um Telefone/Email.
     api.apagaTelefone = (req, res) => {
-        const knex = app.conexao.conexaoBDKnex();
         const { cod_telefone, cod_pessoa } = req.params;
 
-        knex('telefone').where('cod_telefone', cod_telefone).andWhere('cod_pessoa', cod_pessoa).delete()
-            .then(resultado => {
-                knex.destroy();
-                res.status(200).end();
-            })
-            .catch(erro => {
-                console.log(erro);
-                knex.destroy();
-                res.status(500).send(app.api.erroPadrao());
-            });
+        if(cod_telefone && cod_pessoa){
+            const knex = app.conexao.conexaoBDKnex();
+
+            knex('telefone').where('cod_telefone', cod_telefone).andWhere('cod_pessoa', cod_pessoa).delete()
+                .then(resultado => {
+                    knex.destroy();
+                    res.status(200).end();
+                })
+                .catch(erro => {
+                    console.log(erro);
+                    knex.destroy();
+                    res.status(500).send(app.api.erroPadrao());
+                });
+        } else { res.status(400).send(app.api.erroPadrao()); }
     }
 
 
@@ -231,23 +258,26 @@ module.exports = function(app){
 
     //Apaga uma Pessoa.
     api.apagaPessoa = (req, res) => {
-        const knex = app.conexao.conexaoBDKnex();
         const { cod_pessoa } = req.params;
 
-        knex('pessoa').where('cod_pessoa', cod_pessoa).delete()
-            .then(resultado => {
-                knex.destroy();
-                res.status(200).end();
-            })
-            .catch(erro => {
-                console.log(erro);
-                knex.destroy();
-                if(erro.errno == 1451){
-                    res.status(500).send('Esta pessoa não pode ser apagada pois existem outras informações associadas a ela.');
-                } else {
-                    res.status(500).send(app.api.erroPadrao());
-                }
-            });
+        if(cod_pessoa){
+            const knex = app.conexao.conexaoBDKnex();
+
+            knex('pessoa').where('cod_pessoa', cod_pessoa).delete()
+                .then(resultado => {
+                    knex.destroy();
+                    res.status(200).end();
+                })
+                .catch(erro => {
+                    console.log(erro);
+                    knex.destroy();
+                    if(erro.errno == 1451){
+                        res.status(500).send('Esta pessoa não pode ser apagada pois existem outras informações associadas a ela.');
+                    } else {
+                        res.status(500).send(app.api.erroPadrao());
+                    }
+                });
+        } else { res.status(400).send(app.api.erroPadrao()); }
     }
         
     return api;
